changes to support larger range

+++ projects/hipblaslt/library/src/amd_detail/hipblaslt-ext.cpp
@@ -412,7 +412,7 @@ namespace hipblaslt_ext
     public:
         uint16_t splitK = 0;
         int16_t  wgm    = 0;
-        uint16_t  skgrid = 0;
+        uint32_t  skgrid = 0;
     };
 
     GemmTuning::GemmTuning()
@@ -446,7 +446,7 @@ namespace hipblaslt_ext
         pimpl->wgm = wgm;
     }
 
-    void GemmTuning::setSKGrid(uint16_t skgrid)
+    void GemmTuning::setSKGrid(uint32_t skgrid)
     {
         pimpl->skgrid = skgrid;
     }
@@ -461,7 +461,7 @@ namespace hipblaslt_ext
         return pimpl->wgm;
     }
 
-    uint16_t GemmTuning::getSKGrid() const
+    uint32_t GemmTuning::getSKGrid() const
     {
         return pimpl->skgrid;
     }
